fix(parser): do not slice lines in HEREDOC_BLOCK
Originally, the parser sliced these lines since the first & last lines
of a HEREDOC_BLOCK are just the delimiters, which are not meaningful to
the parser or renderer. However, the lexer already handles that, which
means slicing lines from the HEREDOC_BLOCK token in the parser actually
removes meaningful lines. This commit fixes that by not slicing
anything from the HEREDOC_BLOCK in the parser.