feat(lexer): scan meaningful whitespace for tokens
The lexer previously discarded most whitespace between tokens, which led
to some awkward handling of how whitespace was rendered between tokens.
For example, the string "[a]" was lexed as [ BRACKET, WORD ("a]") ]. The
renderer previously had no information about the whitespace between
these tokens and simply rendered them by joining the lexemes with a " "
(space) character. But this resulted in the output "[ a]", which is
incorrect (it should be the same "[a]" as in the source string).
Including the whitespace data in the tokens will allow the renderer to
more accurately render the source.