fix(parser): fix repeated LIST_ITEM nodes
Previously, handleList called handleListItems on the original st that
was passed to handleList. handleListItems then added the LIST_ITEM nodes
to the st it was passed, and returned the final state. After that,
however, handleList took the state and added its AST as the
ITEMIZE/ENUMERATE node's children, which basically ended up nesting the
entire AST into the list and repeating all the nodes. This commit fixes
that by properly taking the sublist containing all the LIST_ITEM tokens
and only passing those in the state to handleList so that the final
state's AST only consists of the list items which can then be added as
children to the ITEMIZE/ENUMERATE node.