fix(parser): check tokens before slicing from list
Previously, the handleListItems function would slice the first and last
tokens from the sublist, where each sublist represented one list item.
This was meant to fix recursion since the first and last tokens would
be UL_ITEM/OL_ITEM tokens which are boundaries for list items. However,
this is not the case for the very last list item. Since there are no
other list items that follow the last one, it has no boundary at the
end; the advanceWhile function stops because it runs out of tokens. In
this case, slicing the last token actually removed a token that was
supposed to be part of the list item's content. This commit fixes that
by ensuring the last item's last token is not sliced off.