fix(parser): check tokens before slicing from para
This is similar to the previous bug with list items. When the
handleParagraphBoundary function is called right after parsing a section
node, the state's current token may not actually be a paragraph boundary
token, since there is no special boundary token that follows section
tokens. Thus, slicing from the start of the sublist in this case
actually gets rid of meaningful tokens, which we do not want. This
commit checks to make sure the token really is a special boundary token
(in this case, just an EMPTY_ROW) before slicing it off.