fix(lexer): check \n before slicing from HEREDOC
Similar bug to the previous slicing ones in the parser. Here, if the
heredoc block appears at the very end of the file, for example, then the
slicing of the string results in the last character of the closing
delimiter being removed (we incorrectly assumed that the last character
would always be a \n). This fixes that by adding a check for the \n
before slicing.