fix(lexer): properly lex heredocs at end of file
The lexer previously assumed a newline \n always followed the closing
delimiter in heredoc blocks, which meant it did not lex properly if it
appeared at the very end of a file (i.e no \n after the closing
delimiter). This commit fixes that by using the appropriate predicates
when advancing.