refactor(parser): rewrite parser
Like the lexer, the parser too avoids recursive function calls (instead,
it uses a stack now). There's still bugs to iron, with how padding is
dealt with and some other special handling for ndoes like HEREDOC_TEX,
but the parser more or less works. Besides the bugs, the code is also
rather messy right now and could do with some refactoring and nice
comments for documentation.