fix(lexer): properly handle empty rightPads
captureRightPad ends up including non-whitespace characters when there
is no right padding after a token, because of captureMarkedSubstring
being inclusive. This commit changes the advance function slightly so
that we skip the first char (i.e last char of the token), stop advancing
when we're ON a whitespace/the end of the rightPad, and then slices off
the last character since we know it's not part of the rightPad. This
ensures extra non-rightPad characters are not included.